DERBY-1787 contributed by Mamta Satoor patch: DERBY1787_UseCorrectTerminologyV1diff.txt
Grant revoke functionality was added in Derby 10.2 The comments that went into the grant revoke code, in some places refer to database owner as "dba". They are not the same thing. In the grant revoke world, dba is a role. We haven't added roles into Derby yet but current use of dba in comments might make it confusing when we do start working on roles including dba.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@448424 13f79535-47bb-0310-9956-ffa450edef68