DERBY-1854 (SYSCS_COMPRESS_TABLE corrupts table with a single column which is both a primary key and a foreign key)
Problem was all the conglomerate descriptor entries in sys.sysconglomerates
were not getting updated with new conglomerate number generated for an index 
during compress/bulk-insert, when an index is shared. Update code was assuming 
conglomerate id is common when an index is shared, but that is not correct. 
ConglomerateID's in sys.sysconglomerates are unique. 

This patch modifies the update conglomerate descriptor code to update each
conglomerate descriptor separately using conglomerateID as the key, when
there are more than one conglomerate descriptor referring to the same 
conglomerate(conglomerate number). 

This patch also adds a test-case to test import into a table that has the same column as a primary key and foreign key,  import was also causing corruption because of the same problem.  See DERBY-1641 , for details. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@448758 13f79535-47bb-0310-9956-ffa450edef68