DERBY-1830 contributed by Sunitha Kambhampati
The testcase added for derby-1660 in optimizerOverrides.sql uses jdbc:default:connection and does not work with wctme5.7 foundation causing the diff mentioned in this jira. This testcase needs to be excluded and should run only in environments that have support for DriverManager.

This patch (derby1830.diff.txt) converts this particular testcase into a junit test.

1. Add a new Junit test in lang suite - VTITest.java
2. Conditionally add the test fixture bulkInsertVtiTest if JDBC.vmSupportsJDBC2() is true.
3. Since the problematic testcase is converted to junit test, remove this test from optimizerOverrides.sql and update the master files for embedded and server frameworks.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@449114 13f79535-47bb-0310-9956-ffa450edef68