DERBY-1751: derbynet/testSecMec.java fails with ShutdownException in DerbyNetClient framework
The attached patch avoids the problem seen in this issue by setting
the console output of the network server to a file. This change in
made to the following files:

* functionTests/tests/jdbc4/TestConnectionMethods.java
* functionTests/tests/derbynet/testSecMec.java
* functionTests/tests/derbynet/dataSourcePermissions_net.java
* junit/NetworkServerTestSetup.java

It also changes DRDAProtocolException so that agent Errors will be
printed to the network server console instead of System.out.

Contributed by Fernanda Pizzorno.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@449671 13f79535-47bb-0310-9956-ffa450edef68