DERBY-1866 contributed by Army Brown merging fix (450155) from trunk to 10.1 branch
Attaching a first patch for this issue, d1866_v1.patch. In short, the problem
was that, when pushing predicates to subqueries beneath UNIONs, the predicates
were always being pushed to the *first* table in the subquery's FROM list,
regardless of whether or not that was actually the correct table. Thus it was
possible to push a predicate down to a base table to which it didn't apply,
thereby leading to an assertion failure in sane mode and an index out of
bounds exception in insane mode.

For details on how this occurred and what the fix is, please refer to the code
comments in the patch. The d1866_v1 patch does the following:

1. Adds logic to ensure scoped predicates are only pushed to the appropriate
base tables.

2. Adds one line to OptimizerImpl to solve the hang that
was occuring for the second query shown in repro.sql.
The problem there was just that one variable was not
being properly reset when beginning a new round of
optimization.

3. Adds some test cases to verify the changes for #1 and #2.

Note that the patch is mostly just explanatory comments for existing and new 
logic, plus the test cases.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@451763 13f79535-47bb-0310-9956-ffa450edef68