DERBY-1489: Provide ALTER TABLE DROP COLUMN functionality
This patch provides support for ALTER TABLE t DROP COLUMN c.

The patch modifies the SQL parser so that it supports statements of the form:

  ALTER TABLE t DROP [COLUMN] c [CASCADE|RESTRICT]

If you don't specify CASCADE or RESTRICT, the default is CASCADE.

If you specify RESTRICT, then the column drop will be rejected if it would
cause a dependent view, trigger, check constraint, unique constraint,
foreign key constraint, or primary key constraint to become invalid.

Currently, column privileges are not properly adjusted when dropping a
column. This is bug DERBY-1909, and for now we simply reject DROP COLUMN
if it is specified when sqlAuthorization is true. When DERBY-1909 is fixed,
the tests in altertableDropColumn.sql should be merged into altertable.sql,
and altertableDropColumn.sql (and .out) should be removed.

This new feature is currently undocumented. DERBY-1926 tracks the documentation
changes necessary to document this feature.

The execution logic for ALTER TABLE DROP COLUMN is in AlterTableConstantAction,
and was not substantially modified by this change. The primary changes to
that existing code were:
 -  to hook RESTRICT processing up to the dependency manager so that
    dependent view processing was sensitive to whether the user
    had specified CASCADE or RESTRICT
 -  to reread the table descriptor from the catalogs after dropping all the
    dependent schema objects and before compressing the table, so that the
    proper scheman information was used during the compress.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@453420 13f79535-47bb-0310-9956-ffa450edef68