DERBY-1939
Adding an ASSERT to catch when the problem happens, rather than later
as a side effect.  The problem is that template created to manipulate
the hash base table gets a null for the HeapRowLocation, later when that
template is used store interprets the null as an object has been stored
on disk and reads it as if it is a self described object rather than
a HeapRowLocation.

This just adds the assert, it does not fix the issue.

Also added some comments that helped me understand what was going on.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@453832 13f79535-47bb-0310-9956-ffa450edef68