DERBY-1939
Fixes HeapRowLocation.getNewNull() to return actual null object, this 
functionality is used by the hash overflow to disk code.  Without this
change the HeapRowLocation is correctly written out to the StoredPage, 
but on read the overflow code passes a null template which causes the
wrong code to read the data back into the object. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@454410 13f79535-47bb-0310-9956-ffa450edef68