DERBY-1785 contributed by Myrna van Lunteren patch: DERBY-1785_20061007.diff
Attaching a band-aid patch for this issue. I chose to comment out the method 
rather than remove as a way to document the quirky behavior.

Having the method setSecurityProps overload the one in jvm.java causes 
problems when running the junit tests - they *do* successfully run with 
securityManager.
Foundation class tests actually run ok with security manager - except when 
useprocess is false. This is caused by a bug in the jvm. See also DERBY-885.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@462607 13f79535-47bb-0310-9956-ffa450edef68