DERBY-183: Allow unnamed parameters in CREATE FUNCTION
This patch was contributed by James F. Adams (derby@xemaps.com)

The patch does the following:

1) Modifies java/engine/org/apache/derby/impl/sql/compile/sqlgrammar.jj
    a) Initializes parameterName to "" in procedureParameterDefinition
       and functionParameterDefinition
    b) Makes parameterName optional in procedureParameterDefinition
       and functionParameterDefinition

2) Modifies java/engine/org/apache/derby/impl/sql/compile/CreateAliasNode.java
   to ignore function and procedure parameter names equal to "" when
   checking for duplicate parameter names. 

Tests have been added to lang/functions.sql and lang/procedure.java. 

The parameter name is made optional by surrounding its production with [].

This changes the grammar from:

parameterName = identifier(Limits.MAX_IDENTIFIER_LENGTH, true)
  typeDescriptor = dataTypeDDL()

to:

        [ parameterName = identifier(Limits.MAX_IDENTIFIER_LENGTH, true) ]
        typeDescriptor = dataTypeDDL()

This results in a choice conflict because certain tokens satisfy both
identifier() and dataTypeDDL(). An additional token of lookahead resolves
this conflict. This results in:

        [ LOOKAHEAD(2) parameterName = identifier(Limits.MAX_IDENTIFIER_LENGTH, true) ]
        typeDescriptor = dataTypeDDL()

Expressing this in an alternate form such as:

        (
                parameterName = identifier(Limits.MAX_IDENTIFIER_LENGTH, true)
                typeDescriptor = dataTypeDDL()
        ) | typeDescriptor = dataTypeDDL()

still results in a choice conflict so I opted for the more compact form.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@463982 13f79535-47bb-0310-9956-ffa450edef68