DERBY-737: SYSCS_UTIL.SYSCS_COMPRESS_TABLE should create statistics
This patch was contributed by Mamta A. Satoor (msatoor@gmail.com)

The changes have been very localized in
AlterTableConstantAction.java!updateIndex(). Currently, this method checks
if statistics already exist for an index. If yes, then it sets a flag
updateStatistics to true. Later, the code checks for this flag and drops
the existing statistics and creates new statistics for that index provided
the user table at this point is not empty. So, as we can see, if there is
an index with no preexisting statistics, the flag updateStatistics will be
set to false and hence no statistics related code is executed and hence
even though the user table is not empty at the time of compress, no
statistics get generated for such an index.

I am proposing to fix the problem by still using the flag to see if
an index has pre-existing statistics. If yes, then we should drop
those statistics. Next, whether the index has pre-existing statistics or
not, go ahead and create new statistics for the index provided the
user table is not currently empty.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@464551 13f79535-47bb-0310-9956-ffa450edef68