DERBY-1240 ( creating /restoring a db from backup using createFrom with log at              different location copies the log from backup to the db dir also).
Problem was restore code in the RawStore was copying all the files that does 
not exist in the database directory after the restore was completed by the 
data factory and log factory boot methods. Because the log is placed at
different location when an external log location is specified, the log
directory does not exist in the database directory. In this case code in the 
RawStore was incorrectly copying log files also because the log directory does 
not exist in the database directory. This patch fixes the problem by restoring
only the files that are not restored by the other factories. Currently, they
are jar files in the database and backup history file.
 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@466221 13f79535-47bb-0310-9956-ffa450edef68