DERBY-1495: Error modifying an identity column after altering the column DERBY-1645: ALTER TABLE SET INCREMENT turns off "Generated By Default"
This patch changes ModifyColumnNode.bindAndValidateDefault so that it
detects the case(s) where the user is altering aspects of an identity column,
and ensures that the other aspects of that identity column are preserved and
not lost. The crucial issue is that if the column is Generated By Default,
then the DefaultInfoImpl column in the SYSCOLUMNS table needs to get
retained when the user uses ALTER TABLE to change either the start value
or the increment value; otherwise the behavior effectively switches from
Generated By Default to Generated Always.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@474502 13f79535-47bb-0310-9956-ffa450edef68