DERBY-1132: Truncation Error with Concat
The Char & Varchar functions were not handled separately to assign to
length for the target type (in the bindExpression() method of
CastNode.java) and thus a default value of 15 was getting assigned to
it (from getColumnDisplaySize() method of DataTypeUtilities class)

Now I added a check for the string types (i.e. both Char & Varchar) in
the source type, and using the operand's maximum length or the maximum
length for that data type (whichever is minimum) and assigning it to
be the length for the target type. Thus avoiding the unwanted
truncation error caused by the default value of 15.

Patch contributed by Saurabh Vyas.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@477168 13f79535-47bb-0310-9956-ffa450edef68