DERBY-1089: NPE inserting a join into a table with a generated column DERBY-2015: NPE in INSERT .. SELECT with IDENTITY column
When an INSERT ... SELECT statement inserts a GENERATED ALWAYS identity
column, the identity column's column reference is NULL, since that
column does not have a corresponding column in the SELECT list. This
NULL expression needs to be skipped over when remapping column references
from the SELECT column list to the INSERT column list. 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@478622 13f79535-47bb-0310-9956-ffa450edef68