DERBY-1204: TRIGGER with multi-row INSERT action throws OutOfBounds exception
This patch was contributed by Yip Ng (yipng168@gmail.com)

The statement fails because the bindReferencesClause() method in
CreateTriggerNode did not filter those column references that are not
relevent to the conversion processing for OLD/NEW transition variables.
The above trigger action will have the following parse tree: (roughly)

              InsertNode
                 |
              UnionNode
               /     \
        UnionNode    RowRSNode
         /    \            |
RowRSNode RowRSNode  NumericConstantNode
      |       |                    |
 ColRef    NumericConstantNode    333
      |       |
     n.i     999

Note that the CollectNodesVisitor collects *ALL* the ColumnReferences in
the trigger action. In the above case, it returned 3 column references
nodes. One in the top level Union node, one in top level Union Node's
left child which it is also an UnionNode and the last one is found in
the RowResultSetNode.

The UnionNodes will have a copy of the CRs from its left child and
those CRs will not have its beginOffset set which indicates they are
not relevant for the conversion processing here, so the corrective
action here is to skip these entries.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@482358 13f79535-47bb-0310-9956-ffa450edef68