DERBY-2202: DROP PROCEDURE depends on SET SCHEMA
This patch was contributed by Yip Ng (yipng168@gmail.com)

The symptom of this jira is the same as those described in DERBY-1304
for DROP VIEW except that it applies to DROP PROCEDURE, DROP FUNCTION
and DROP SYNONYM statements. (Their bind logic are centralized in
DropAliasNode.) The DropAliasNode's bindStatement() should have created
a dependency on the statement, so when its associated alias descriptor
gets dropped, the statement can be invalidated accordingly. This explains
why subsequent DROP PROCEDURE statement fails.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@491632 13f79535-47bb-0310-9956-ffa450edef68