DERBY-2195: Trigger problems after maximum trigger count exception
This patch was contributed by Yip Ng (yipng168@gmail.com)

The system is not cleaning up the statement context properly. The
cleanup should have been performed in GenericTriggerExecutor's executeSPS
exception handler. The GenericTriggerExecutor is responsible to make sure
the SPS execution is cleanup correctly.

In addition, there is another potential problem with the exception
handling logic in executeSPS() where it assumes that the SPS has executed
and therefore will have a new statement context and pops it off from
the LCC. This is somewhat dangerous as StandardException can occur before
pushing the new statement context to the stack; hence, this might
result in popping the wrong statement context. So the patch
also addresses this issue.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@492741 13f79535-47bb-0310-9956-ffa450edef68