DERBY-2183: Trigger recompilation problem when not qualified with schema
backporting fix from trunk to 10.1 branch.

This patch was contributed by Yip Ng (yipng168@gmail.com)

The problem was that the compilation schema for the SPS was never set
during its recompilation, it was using the current schema to perform the
recompilation which is not correct. It should have use the SPS's
compilation schema to do the recompilation when we push the compiler context.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@493253 13f79535-47bb-0310-9956-ffa450edef68