DERBY-2152: Clean-up patch that does the following:
  1. Creates a new "init()" method for NewInvocationNode that is specifically
targeted for mapping VTI "tables" and VTI "table functions" to their
corresponding class names. The new init() method takes both a TableDescriptor
and a TableName, exactly one of which must be null. If the TableDescriptor
is null then we will resolve the TableName to a VTI "table function"; if the
TableName is null then we will resolve the TableDescriptor to a VTI "table".
Thus VTI classes are now consistently resolved inside the NewInvocationNode
class.

  2. Updates NodeFactory.mapTableAsVTI() to make use of the new "init()"
method in NewInvocationNode. mapTableAsVTI() used to take as a parameter the
resolved VTI class name; now it just passes the received TableDescriptor to
the new init() method in NewInvocationNode and the latter resolves the VTI
class.

  3. Combines the getVTIClassForTable() and getVTIClassForTableFunction()
methods of DataDictionary into a new method, "getVTIClass()", that takes a
TableDescriptor and a boolean value. The boolean indicates the kind of
mapping to do--"table" or "table function"--while the TableDescriptor holds
the schema and name information used for resolving the class. This resolves
the inconsistency between the argument styles used in the old getVTIClassFor*()
methods.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@494993 13f79535-47bb-0310-9956-ffa450edef68