DERBY-2191: Cleanup of FormatableBitSet
Changed the bitset operator methods or(), and() and xor() so that they
follow the same pattern. All are now performing the operation
bytewise, and there is no special handling of the last partial
byte. The patch also adds a method called invariantHolds() that checks
if the class' invariant is maintained (for use in the unit test).

Patch contributed by Dyre Tjeldvoll.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@498772 13f79535-47bb-0310-9956-ffa450edef68