DERBY-1909: ALTER TABLE DROP COLUMN needs to update GRANTed privileges
When ALTER TABLE DROP COLUMN is used to drop a column from a table, it needs to update the GRANTed column privileges on that table.

The core of this proposed patch involves refactoring and reusing the
DERBY-1847 method which knows how to rewrite SYSCOLPERMS rows
to update the COLUMNS column. The DERBY-1847 version of that code
only handled the case of adding a bit to the COLUMNS column; this patch
extends that method to support removing a bit from the COLUMNS
column as well, then calls the method from the AlterTable execution logic.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@503550 13f79535-47bb-0310-9956-ffa450edef68